Update constants: non-confusing parts.
Regenerate constants via tool-script, but only commit
the "simple" cases: i.e. those that don't have
conflicting duplicate names.
For the complicated cases, we'll need to improve
the disambiguation code first.

This cleans up a lot of leftover non-breaking spaces.
These used to be in the upstream spec HTML document,
but were removed at some point in the past, according
to a deep dive in the Internet Archive Wayback Machine.